feat(gasEstimations): add new simulation contracts for gas estimation
