+++ src/rpc/estimation/utils.ts
@@ -1,0 +1,34 @@
+import { type Hex, decodeErrorResult, parseAbi } from "viem"
+
+export function parseFailedOpWithRevert(data: Hex) {
+    try {
+        const decoded = decodeErrorResult({
+            abi: parseAbi(["error Error(string)", "error Panic(uint256)"]),
+            data
+        })
+
+        if (decoded.errorName === "Error") {
+            return decoded.args[0]
+        }
+
+        if (decoded.errorName === "Panic") {
+            // from https://docs.soliditylang.org/en/v0.8.0/control-structures.html
+            const panicCodes: { [key: number]: string } = {
+                1: "assert(false)",
+                17: "arithmetic overflow/underflow",
+                18: "divide by zero",
+                33: "invalid enum value",
+                34: "storage byte array that is incorrectly encoded",
+                49: ".pop() on an empty array.",
+                50: "array sout-of-bounds or negative index",
+                65: "memory overflow",
+                81: "zero-initialized variable of internal function type"
+            }
+
+            const [code] = decoded.args
+            return panicCodes[Number(code)] ?? `${code}`
+        }
+    } catch {}
+
+    return data
+}
